Refatoracao para funcoes e arquivos
