+++ src/main/kotlin/calculadora/core/Resultados.kt
@@ -1,0 +1,15 @@
+package calculadora.core
+
+import java.time.LocalDateTime
+
+private var resultados: Array<String> = emptyArray()
+
+fun salvarResultado(resultado: String) {
+    resultados += "${LocalDateTime.now()}: $resultado"
+}
+
+fun listarResultados() {
+    for (resultado in resultados.reversed()) {
+        println(resultado)
+    }
+}
